не получается отловить причину ошибки:
[2017-02-14T18:53:06.693+0300] [WARNING] [] [javax.enterprise.web] [tid: _ThreadID=28 _ThreadName=http-thread-pool(4)] [timeMillis: 1487087586693] [levelValue: 900] [[
  StandardWrapperValve[ru.dcherneta.vcpe.lib.db.service.RestApplication]: Servlet.service() for servlet ru.dcherneta.vcpe.lib.db.service.RestApplication threw exception
java.lang.IllegalArgumentException: object is not an instance of declaring class
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)

+++ lib/src/main/java/ru/dcherneta/vcpe/lib/db/service/ItemRESTService.java
@@ -25,7 +25,7 @@ public class ItemRESTService implements EntityRESTService<ItemEntity>{
     @Override
     @GET
     //@Produces(MediaType.APPLICATION_JSON)
-    public ItemEntity get(@QueryParam("id") int id){
+    public ItemEntity get(@QueryParam("id") Long id){
         return entityManager.find(ItemEntity.class, id);
     }
     /*
